refactor: separate sqlite references from workflow indexes
