+++ tests/indexes/test_api.py
@@ -14,9 +14,9 @@ from tests.fixtures.client import ClientSpawner, JobClientSpawner
 from tests.fixtures.response import RespIs
 from virtool.fake.next import DataFaker
 from virtool.history.sql import SQLLegacyHistory
-from virtool.indexes.constants import INDEX_SQLITE_FILE_NAME
 from virtool.indexes.sql import SQLIndex, SQLIndexFile
 from virtool.indexes.utils import compose_index_file_key
+from virtool.references.sqlite import REFERENCE_SQLITE_FILE_NAME
 from virtool.storage.protocol import StorageBackend
 from virtool.workflow.pytest_plugin.utils import StaticTime
 
@@ -280,7 +280,7 @@ async def test_download(
             assert await response.read() == expected_bytes
 
 
-async def _seed_downloadable_sqlite(
+async def _seed_downloadable_sqlite_reference(
     fake: DataFaker,
     memory_storage: StorageBackend,
     pg: AsyncEngine,
@@ -288,7 +288,7 @@ async def _seed_downloadable_sqlite(
     user = await fake.users.create()
     reference = await fake.references.create(user=user)
     index = await fake.indexes.create(reference, user)
-    expected = b"server-produced SQLite index"
+    expected = b"server-produced SQLite reference"
 
     async with AsyncSession(pg) as session:
         storage_key = await session.scalar(
@@ -297,7 +297,7 @@ async def _seed_downloadable_sqlite(
 
         session.add(
             SQLIndexFile(
-                name=INDEX_SQLITE_FILE_NAME,
+                name=REFERENCE_SQLITE_FILE_NAME,
                 index=str(index.id),
                 index_id=index.id,
                 type="sqlite",
@@ -310,40 +310,52 @@ async def _seed_downloadable_sqlite(
         yield expected
 
     await memory_storage.write(
-        compose_index_file_key(storage_key, INDEX_SQLITE_FILE_NAME),
+        compose_index_file_key(storage_key, REFERENCE_SQLITE_FILE_NAME),
         stream(),
     )
 
     return index.id, expected
 
 
-async def test_download_sqlite_for_jobs(
+async def test_download_sqlite_reference_for_jobs(
     fake: DataFaker,
     memory_storage: StorageBackend,
     pg: AsyncEngine,
     spawn_job_client: JobClientSpawner,
 ):
-    """The jobs download route serves a recorded SQLite index file."""
+    """The jobs download route serves a recorded SQLite reference file."""
     client = await spawn_job_client(authenticated=True)
-    index_id, expected = await _seed_downloadable_sqlite(fake, memory_storage, pg)
+    index_id, expected = await _seed_downloadable_sqlite_reference(
+        fake,
+        memory_storage,
+        pg,
+    )
 
-    response = await client.get(f"/indexes/{index_id}/files/{INDEX_SQLITE_FILE_NAME}")
+    response = await client.get(
+        f"/indexes/{index_id}/files/{REFERENCE_SQLITE_FILE_NAME}"
+    )
 
     assert response.status == HTTPStatus.OK
     assert await response.read() == expected
 
 
-async def test_download_sqlite_for_authenticated_user(
+async def test_download_sqlite_reference_for_authenticated_user(
     fake: DataFaker,
     memory_storage: StorageBackend,
     pg: AsyncEngine,
     spawn_client: ClientSpawner,
 ):
-    """The authenticated route serves a recorded SQLite index file."""
+    """The authenticated route serves a recorded SQLite reference file."""
     client = await spawn_client(authenticated=True, administrator=True)
-    index_id, expected = await _seed_downloadable_sqlite(fake, memory_storage, pg)
+    index_id, expected = await _seed_downloadable_sqlite_reference(
+        fake,
+        memory_storage,
+        pg,
+    )
 
-    response = await client.get(f"/indexes/{index_id}/files/{INDEX_SQLITE_FILE_NAME}")
+    response = await client.get(
+        f"/indexes/{index_id}/files/{REFERENCE_SQLITE_FILE_NAME}"
+    )
 
     assert response.status == HTTPStatus.OK
     assert await response.read() == expected
