+++ tests/indexes/test_db.py
@@ -8,7 +8,6 @@ from sqlalchemy.ext.asyncio import AsyncEngine, AsyncSession
 import virtool.indexes.db
 from virtool.fake.next import DataFaker
 from virtool.history.sql import SQLLegacyHistory
-from virtool.indexes.constants import INDEX_SQLITE_FILE_NAME
 from virtool.indexes.db import (
     INDEX_FILE_NAMES,
     JOB_INDEX_FILE_NAMES,
@@ -20,13 +19,14 @@ from virtool.indexes.db import (
 )
 from virtool.indexes.sql import SQLIndex, SQLIndexFile
 from virtool.otus.sql import SQLOTU
+from virtool.references.sqlite import REFERENCE_SQLITE_FILE_NAME
 from virtool.utils import timestamp
 
 
-def test_sqlite_file_contract():
+def test_sqlite_reference_file_contract():
     """SQLite is downloadable but not classified as a legacy job artifact."""
-    assert INDEX_SQLITE_FILE_NAME in INDEX_FILE_NAMES
-    assert INDEX_SQLITE_FILE_NAME not in JOB_INDEX_FILE_NAMES
+    assert REFERENCE_SQLITE_FILE_NAME in INDEX_FILE_NAMES
+    assert REFERENCE_SQLITE_FILE_NAME not in JOB_INDEX_FILE_NAMES
 
 
 async def _seed_index(pg: AsyncEngine, fake: DataFaker, legacy_id: str) -> int:
