--- virtool/indexes/constants.py
@@ -1,3 +1,0 @@
-"""Constants for index file contracts."""
-
-INDEX_SQLITE_FILE_NAME = "virtool-index-sqlite.v1.sqlite"
